requirements.txt
pandas
seaborn
matplotlib
scikit-learn
statsmodels
streamlit
plotly
numpy
pyyaml
